Using common class for inputs
